Print in F280 only when asked, handle file in the class
Printing from counting is now optional (set in code).
Printing of F280 can be requested using - (dash).

F280 file output is now handled in the F280 object.
Same file is being opened and used in case of multiple simulations
and unline the previous implementation, content is now overwritten
not accumulated which did not make much sense.