Add clustered pest for stems
Uses beta distribution to determine number of stems which should have pest
(for all shipments, i.e., unless this is zero, all are infested to some extent).

If cluster is too big in terms of number of stems, smaller clusters with max cluster size are
created plus a cluster with the remaining stems.

If the cluster is too spread out, i.e., outside of the list of stems, it is rescaled
into the range.

Clusters can overlap, but only presence (or absence) is tracked. Same applies to cluster
containing one stem (index) more than once. Consequently, the resulting number infested
stems is lower than the originaly requested.