Add continuous and uniform random distributions for clusters (#6)
The distributions produce exact number of stems requested for a cluster.

Adds a test for infestation rate of basic types of clusters.
This currently can't be tested in the code directly because
gamma and small max_stems_per_cluster causes small infestation rates.
Max cluster size is large to make clear it is not reached.
A cluster needs to accommodate at least all infested stems per cluster.

Uses a fixed seed to ensure consistent results. The test fails (with this seed)
for earlier (wrong versions) of this code which don't limit size of cluster
also by number of stems
because the size of cluster should be limited not only by the settings but also by the
number of available stems.