Add inspection strategies (#12)
This adds new strategies and changes most of the inspection code to accommodate the additions.   

Sampling strategy can operate on stems or boxes and new hypergeometric and percentage functions handle the number of units to inspect.

Strategy for selection of individual boxes to inspect is a new separate concept.

Boxes are inspected to a percentage given in the configuration. On the box level, the all selected boxes are sampled both to completion and to detection, i.e., the stats are collected for both options without user choosing one or another.