Use SimpleNamespace to make return value less error prone (#18)
types.SimpleNamespace is now used for return value so that retrieving the values does not require the caller to get the order right. The assignments are now more explicit and potentially more readable.

This also reduces number of variables in the inspect() function thus fixing Flake8 R0914: Too many local variables (19/15) from #17.