Create two examples for the notebook
Replace the current example by two examples where the configuration is based on
the sample config.yml in the repo, but parts not needed are removed.
One example is extra small and simplified just to show the individual stems, boxes, and pests.
The other example is more realistic and more similar to the current config.yml.