Split doc into multiple files (#52)
Remove documentation from the readme. Use separate files instead of one large file.
Link the documentation and resources in the readme.

Remove end_strategy from doc and other files. Leave explanation of the two strategies because both are applied. This could be part of explanation of outputs in the future. 

This also: Links doc pages using Next. Adds page about Binder and pip dependencies for developers into the readme.