Introduce Shipment class to allow attribute access (#66)
Newly introduced Shipment supports both item access as dictionary and attribute access which is useful
when there is a lot of variables around or variables are used just to avoid the more complicated dictionary syntax.

Pylint warns about too many ancestors (8) for the class, but that's on the library side, so not relevant. The warning was thus disabled. Black requires an empty line between docstring and a comment (further dividing documentation from the implementation details - disabled warnings - which is fine).