Redo configuration of shipments
Now everything is nested under shipment key, the input F280 file has a better name,
stems per box numbers handleding is fixed, and everything is under one shipment key
which is the case for rest of the options. Draft of documentation included. Some tests updated.

This is WIP and it does not include changes which would conflict with #65
(consequently, all tests are failing). This needs to be finished after #65 is in.