Add export function to pandas and example notebook
The library is extended only by one function exporting to pandas DataFrame, now mostly (unfortunate) duplicate of CSV export.

A notebook is included which demonstrates the use and adds couple of examples with pandas, plotting, and seaborn.

A simple test is added too just to ensure things are running, but does not perform any checks.