Add doc for running the simulation
This adds documentation for running the simulation. It is mostly the intro and the other files are referenced.

The new page goes in between config of contamination and outputs in terms of the 'Next:' link.

Headings in the use cases page are now fixed and the synthetic F280 page is linked from there
(all pages should be now linked from somewhere).