Add doc for running the simulation (#113)
This adds documentation for running the simulation. It is mostly the intro and the other files are referenced.

The new page goes in between config of contamination and outputs in terms of the 'Next:' link.

Headings in the use cases page are now fixed and the synthetic F280 page is linked from there
(all pages should be now linked from somewhere).

Updates use cases. Reordered these by what I think will be most interesting. Added one about slippage rates.
Removed one that focused on CFRP and might be confusing now when it is not fully implemented.

Co-authored-by: kellynm <32572763+kellynm@users.noreply.github.com>