Test more seeds in basic simulation run
Not all seeds generate error which the code already accounted for.
However, 10 seeds was not enough to capture #123 which needed seed=15.
Test with seeds starting at 0, 100, 200, ..., 9000 showed that the #123 issue is hit
at first 10 seeds in 8 cases and in first 20 in two cases., so using 30 seeds
in the test to capture #123 plus 10 extra seeds, but still limit the number to minimum.
(Not using random seeds (which is what how #123 was actually discovered) because
that may result in random, hard to reproduce, errors in CI, so the usefulness is limited.)