Add Cut Flower Release Program
This is a draft which includes basic structure for CFRP and moves all functionality to determine if the inspection is needed to a separate file.
The new module is called skipping as in skipping inspection, but it could be also release as in release programs (to keep it one word only). The config key is still called release_programs. One issue is that release happens even outside of release programs.
New tests are testing CFRP code directly.

What is missing is reading of CFRP schedule (for tests, it is provided directly in the code).