Use f-strings, fix pretty style error message
The strings with format and locals() were meant to be replaced by f-strings when minimal Python version is high enough. This makes the replacement in most of the code. Additionally, Pylint now reports C0209: Formatting a regular string which could be a f-string (consider-using-f-string) and this fixes that warning for most places.

Some more complex cases are not addressed here.

This discoveres and fixes an error in the pretty consignment style error message which was not previosly visible to the Flake8.