Use f-strings, fix pretty style error message (#154)
The strings with format and locals() were meant to be replaced by f-strings when minimal Python version is high enough. This makes the replacement in most of the code. Additionally, Pylint now reports C0209: Formatting a regular string which could be a f-string (consider-using-f-string) and this fixes that warning for most places.

Some more complex cases are not addressed here.

This discovers and fixes an error in the pretty consignment style error message which was not previously visible to the Flake8.

Actually create variable with a complicated value.

Consignment has now attributes which is simpler syntax than dict access.