Add Cut Flower Release Program (#151)
This includes CFRP and moves all skip functionality (determine if the inspection is needed) to a separate file. The new module is called skipping as in skipping inspection (or skip lot). It is not release as in release programs because that seems misleading and not general enough (whether or not we keep it to one word only).

The config key is still called release_programs. One issue is that release happens even outside of release programs.

Reads CFRP schedule from a CSV file with all combinations of date, origin, and flower. Internally, it is represented as a dictionary with origin and flower as a key and set of dates as a value which is syntactically easy and reasonably fast lookup.

The CFRP has also a list of participating ports (other ports always inspected). It is one list of ports for the whole program.

Contains user documentation and tests.