Document effectiveness, improve code and tests
This adds documentation for inspection effectiveness. It modifies the description of tolerance level just to be clear what is the difference.

The implementation is modified to throw an exception instead of printing when effectiveness is wrong and otherwise to match how parameters are treated elsewhere.

Tests are using fixtures and parametrization instead of loops and global variables mixed with a class fixture.