Add single-parameter clustering with contaminated subset and rate (#210)
This adds a new function for item contamination (not for boxes) which uses one cluster which is a subset based on clustering parameter and contamination rate.

This includes documentation, tests, and experiment in a notebook. (It recomputes the other experimental part of the clustering notebook.)

It reuses the `clustered/distribution` parameter used in the more complex clustering methods and is enabled when `clustered/distribution` is set to `subset`. In the documentation, the old clustering is now called _multi-parameter clustering_ and the new one is called _single-parameter clustering_ because one parameter is driving the clustering itself.

Also update from deprecated NumPy array to scalar conversion in the notebook.