Add parallel execution of scenarios (#219)
Creates a variant of scenarios run function which also takes a single config and a scenario table, but executes the scenarios in parallel using concurrent.futures.

This adds an option to save the individual result from each scenario instead of just the totals and also saves the seed which is associated with each result.